+++ NEWS.md
@@ -2,8 +2,12 @@ This document attempts to list user-visible changes and any major internal
 rearrangements of Notcurses.
 
 * 3.0.1 (not yet released)
-  * Added `ncplane_growtext()`, which allows you to dump text to a plane
-    (ala `ncplane_puttext()`), and have it grow right along with you.
+  * Added the `NCPLANE_OPTION_VSCROLL` flag. Creating an `ncplane` with this
+    flag is equivalent to immediately calling `ncplane_set_scrolling(true)`.
+  * Added the `NCPLANE_OPTION_AUTOGROW` flag and the `ncplane_set_autogrow()`
+    and `ncplane_autogrow_p()` functions. When autogrow is enabled, the plane
+    is automatically enlarged to accommodate output at its right (no scrolling)
+    or bottom (scrolling enabled) boundaries.
 
 * 3.0.0 (2021-12-01) **"In the A"**
   * Made the ABI/API changes that have been planned/collected during 2.x